+++ config/application.rb
@@ -23,12 +23,11 @@ module Registry
 
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
-    config.time_zone = ENV['time_zone'] || 'Tallinn'  # NB! It should be defined, 
+    config.time_zone = ENV['time_zone'] || 'Tallinn'  # NB! It should be defined,
                                                       # otherwise ActiveRecord usese other class internally.
 
-    # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
-    # config.i18n.default_locale = :de
+    config.i18n.load_path += Dir[Rails.root.join('config', 'locales', '**', '*.yml').to_s]
+    config.i18n.default_locale = :en
 
     config.paths.add File.join('app', 'api'), glob: File.join('**', '*.rb')
     config.autoload_paths += Dir[Rails.root.join('app', 'api', '*')]
@@ -36,6 +35,7 @@ module Registry
     # Autoload all model subdirs
     config.autoload_paths += Dir[Rails.root.join('app', 'models', '**/')]
     config.autoload_paths << Rails.root.join('lib')
+    config.eager_load_paths << config.root.join('lib', 'validators')
 
     # Add the fonts path
     config.assets.paths << Rails.root.join('vendor', 'assets', 'fonts')
@@ -59,6 +59,10 @@ module Registry
       g.javascripts false
       g.helper false
     end
+
+    registrant_portal_uri = URI.parse(ENV['registrant_url'])
+    config.action_mailer.default_url_options = { host: registrant_portal_uri.host,
+                                                 protocol: registrant_portal_uri.scheme }
 
     config.action_mailer.delivery_method = :smtp
     config.action_mailer.perform_deliveries = true