+++ config/initializers/digidoc.rb
@@ -1,0 +1,1 @@
+Digidoc::Client.logger = Rails.application.config.logger