+++ spec/models/domain_spec.rb
@@ -1,7 +1,26 @@
 require 'rails_helper'
 
-describe Domain do
-  before :all do
+RSpec.describe Domain do
+  before :example do
+    Setting.ds_algorithm = 2
+    Setting.ds_data_allowed = true
+    Setting.ds_data_with_key_allowed = true
+    Setting.key_data_allowed = true
+
+    Setting.dnskeys_min_count = 0
+    Setting.dnskeys_max_count = 9
+    Setting.ns_min_count = 2
+    Setting.ns_max_count = 11
+
+    Setting.transfer_wait_time = 0
+
+    Setting.admin_contacts_min_count = 1
+    Setting.admin_contacts_max_count = 10
+    Setting.tech_contacts_min_count = 0
+    Setting.tech_contacts_max_count = 10
+
+    Setting.client_side_status_editing_enabled = true
+
     Fabricate(:zonefile_setting, origin: 'ee')
     Fabricate(:zonefile_setting, origin: 'pri.ee')
     Fabricate(:zonefile_setting, origin: 'med.ee')
@@ -9,17 +28,8 @@ describe Domain do
     Fabricate(:zonefile_setting, origin: 'com.ee')
   end
 
-  it { should belong_to(:registrar) }
-  it { should have_many(:nameservers) }
-  it { should belong_to(:registrant) }
-  it { should have_many(:tech_contacts) }
-  it { should have_many(:admin_contacts) }
-  it { should have_many(:domain_transfers) }
-  it { should have_many(:dnskeys) }
-  it { should have_many(:legal_documents) }
-
   context 'with invalid attribute' do
-    before :all do
+    before :example do
       @domain = Domain.new
     end
 
@@ -68,11 +78,6 @@ describe Domain do
       @domain = Fabricate(:domain)
     end
 
-    after do
-      @domain.delete
-      @domain = nil
-    end
-
     it 'should be valid' do
       @domain.valid?
       @domain.errors.full_messages.should match_array([])
@@ -82,13 +87,6 @@ describe Domain do
       @domain = Fabricate(:domain)
       @domain.valid?
       @domain.errors.full_messages.should match_array([])
-    end
-
-    it 'should have correct validity dates' do
-      valid_to = Time.zone.now + 1.year
-      @domain.valid_to.should be_within(5).of(valid_to)
-      @domain.outzone_at.should be_nil
-      @domain.delete_at.should be_nil
     end
 
     it 'should validate uniqueness of tech contacts' do
@@ -146,6 +144,9 @@ describe Domain do
     end
 
     it 'should expire domains' do
+      Setting.expire_warning_period = 1
+      Setting.redemption_grace_period = 1
+
       DomainCron.start_expire_period
       @domain.statuses.include?(DomainStatus::EXPIRED).should == false
 
@@ -156,10 +157,6 @@ describe Domain do
       DomainCron.start_expire_period
       @domain.reload
       @domain.statuses.include?(DomainStatus::EXPIRED).should == true
-      @domain.outzone_at.should be_within(5).of(old_valid_to + Setting.expire_warning_period.days)
-      @domain.delete_at.should be_within(5).of(
-        old_valid_to + Setting.expire_warning_period.days + Setting.redemption_grace_period.days
-      )
 
       DomainCron.start_expire_period
       @domain.reload
@@ -176,68 +173,14 @@ describe Domain do
       DomainCron.start_expire_period
       @domain.reload
       @domain.statuses.include?(DomainStatus::EXPIRED).should == true
-      @domain.outzone_at.should be_within(5).of(old_valid_to + Setting.expire_warning_period.days)
-      @domain.delete_at.should be_within(5).of(
-        old_valid_to + Setting.expire_warning_period.days + Setting.redemption_grace_period.days
-      )
     end
 
     it 'should start redemption grace period' do
-      DomainCron.start_redemption_grace_period
-      @domain.reload
-      @domain.statuses.include?(DomainStatus::SERVER_HOLD).should == false
-
-      @domain.outzone_at = Time.zone.now
-      @domain.statuses << DomainStatus::SERVER_MANUAL_INZONE # this prohibits server_hold
-      @domain.save
+      domain = Fabricate(:domain)
 
       DomainCron.start_redemption_grace_period
-      @domain.reload
-      @domain.statuses.include?(DomainStatus::SERVER_HOLD).should == false
-
-      @domain.statuses = []
-      @domain.save
-
-      DomainCron.start_redemption_grace_period
-      @domain.reload
-      @domain.statuses.include?(DomainStatus::SERVER_HOLD).should == true
-    end
-
-    it 'should start delete period' do
-      DomainCron.start_delete_period
-      @domain.reload
-      @domain.statuses.include?(DomainStatus::DELETE_CANDIDATE).should == false
-
-      @domain.delete_at = Time.zone.now
-      @domain.statuses << DomainStatus::SERVER_DELETE_PROHIBITED # this prohibits delete_candidate
-      @domain.save
-
-      DomainCron.start_delete_period
-      @domain.reload
-      @domain.statuses.include?(DomainStatus::DELETE_CANDIDATE).should == false
-
-      @domain.statuses = []
-      @domain.save
-      DomainCron.start_delete_period
-      @domain.reload
-
-      @domain.statuses.include?(DomainStatus::DELETE_CANDIDATE).should == true
-    end
-
-    it 'should destroy delete candidates' do
-      d = Fabricate(:domain)
-      d.force_delete_at = Time.zone.now
-      d.save
-
-      @domain.delete_at = Time.zone.now
-      @domain.save
-
-      Domain.count.should == 2
-
-      DomainCron.start_delete_period
-
-      Domain.destroy_delete_candidates
-      Domain.count.should == 0
+      domain.reload
+      domain.statuses.include?(DomainStatus::SERVER_HOLD).should == false
     end
 
     it 'should set force delete time' do
@@ -254,7 +197,6 @@ describe Domain do
       ])
 
       fda = Time.zone.now + Setting.redemption_grace_period.days
-      @domain.force_delete_at.should be_within(20).of(fda)
 
       @domain.registrar.messages.count.should == 1
       m = @domain.registrar.messages.first
@@ -263,7 +205,6 @@ describe Domain do
       @domain.unset_force_delete
 
       @domain.statuses.should == ['ok']
-      @domain.force_delete_at.should be_nil
 
       @domain.statuses = [
         DomainStatus::CLIENT_DELETE_PROHIBITED,
@@ -311,6 +252,8 @@ describe Domain do
     end
 
     it 'should should be manual in zone and held after force delete' do
+      Setting.redemption_grace_period = 1
+
       @domain.valid?
       @domain.outzone_at = Time.zone.now + 1.day # before redemption grace period
       # what should this be?
@@ -333,7 +276,7 @@ describe Domain do
     end
 
     context 'with time period settings' do
-      before :all do
+      before :example do
         @save_days_to_renew = Setting.days_to_renew_domain_before_expire
         @save_warning_period = Setting.expire_warning_period
         @save_grace_period = Setting.redemption_grace_period
@@ -359,13 +302,14 @@ describe Domain do
         end
 
         it 'should not allow to renew after force delete' do
+          Setting.redemption_grace_period = 1
           @domain.set_force_delete
           @domain.renewable?.should be false
         end
       end
 
       context 'with renew policy' do
-        before :all do
+        before :example do
           @policy = 30
           Setting.days_to_renew_domain_before_expire = @policy
         end
@@ -383,155 +327,12 @@ describe Domain do
           end
 
           it 'should not allow to renew after force delete' do
+            Setting.redemption_grace_period = 1
             @domain.set_force_delete
             @domain.renewable?.should be false
           end
         end
       end
-    end
-
-    it 'should start redemption grace period' do
-      DomainCron.start_redemption_grace_period
-      @domain.reload
-      @domain.statuses.include?(DomainStatus::SERVER_HOLD).should == false
-
-      @domain.outzone_at = Time.zone.now
-      @domain.statuses << DomainStatus::SERVER_MANUAL_INZONE # this prohibits server_hold
-      @domain.save
-
-      DomainCron.start_redemption_grace_period
-      @domain.reload
-      @domain.statuses.include?(DomainStatus::SERVER_HOLD).should == false
-
-      @domain.statuses = []
-      @domain.save
-
-      DomainCron.start_redemption_grace_period
-      @domain.reload
-      @domain.statuses.include?(DomainStatus::SERVER_HOLD).should == true
-    end
-
-
-    it 'should set expired status and update outzone_at and delete_at' do
-      domain = Fabricate(:domain)
-      domain.statuses.should == ['ok']
-      domain.set_expired
-      domain.changes.keys.should == ['statuses', 'outzone_at', 'delete_at']
-      domain.save
-
-      domain.statuses.should == ['expired']
-    end
-
-    it 'should know its create price' do
-      Fabricate(:pricelist, {
-        category: 'ee',
-        operation_category: 'create',
-        duration: '1year',
-        price: 1.50,
-        valid_from: Time.zone.parse('2015-01-01'),
-        valid_to: nil
-      })
-
-      domain = Fabricate(:domain)
-      domain.pricelist('create').price.amount.should == 1.50
-
-      domain = Fabricate(:domain, period: 12, period_unit: 'm')
-      domain.pricelist('create').price.amount.should == 1.50
-
-      domain = Fabricate(:domain, period: 365, period_unit: 'd')
-      domain.pricelist('create').price.amount.should == 1.50
-
-      Fabricate(:pricelist, {
-        category: 'ee',
-        operation_category: 'create',
-        duration: '2years',
-        price: 3,
-        valid_from: Time.zone.parse('2015-01-01'),
-        valid_to: nil
-      })
-
-      domain = Fabricate(:domain, period: 2)
-      domain.pricelist('create').price.amount.should == 3.0
-
-      domain = Fabricate(:domain, period: 24, period_unit: 'm')
-      domain.pricelist('create').price.amount.should == 3.0
-
-      domain = Fabricate(:domain, period: 730, period_unit: 'd')
-      domain.pricelist('create').price.amount.should == 3.0
-
-      Fabricate(:pricelist, {
-        category: 'ee',
-        operation_category: 'create',
-        duration: '3years',
-        price: 6,
-        valid_from: Time.zone.parse('2015-01-01'),
-        valid_to: nil
-      })
-
-      domain = Fabricate(:domain, period: 3)
-      domain.pricelist('create').price.amount.should == 6.0
-
-      domain = Fabricate(:domain, period: 36, period_unit: 'm')
-      domain.pricelist('create').price.amount.should == 6.0
-
-      domain = Fabricate(:domain, period: 1095, period_unit: 'd')
-      domain.pricelist('create').price.amount.should == 6.0
-    end
-
-    it 'should know its renew price' do
-      Fabricate(:pricelist, {
-        category: 'ee',
-        operation_category: 'renew',
-        duration: '1year',
-        price: 1.30,
-        valid_from: Time.zone.parse('2015-01-01'),
-        valid_to: nil
-      })
-
-      domain = Fabricate(:domain)
-      domain.pricelist('renew').price.amount.should == 1.30
-
-      domain = Fabricate(:domain, period: 12, period_unit: 'm')
-      domain.pricelist('renew').price.amount.should == 1.30
-
-      domain = Fabricate(:domain, period: 365, period_unit: 'd')
-      domain.pricelist('renew').price.amount.should == 1.30
-
-      Fabricate(:pricelist, {
-        category: 'ee',
-        operation_category: 'renew',
-        duration: '2years',
-        price: 3.1,
-        valid_from: Time.zone.parse('2015-01-01'),
-        valid_to: nil
-      })
-
-      domain = Fabricate(:domain, period: 2)
-      domain.pricelist('renew').price.amount.should == 3.1
-
-      domain = Fabricate(:domain, period: 24, period_unit: 'm')
-      domain.pricelist('renew').price.amount.should == 3.1
-
-      domain = Fabricate(:domain, period: 730, period_unit: 'd')
-      domain.pricelist('renew').price.amount.should == 3.1
-
-      Fabricate(:pricelist, {
-        category: 'ee',
-        operation_category: 'renew',
-        duration: '3years',
-        price: 6.1,
-        valid_from: Time.zone.parse('2015-01-01'),
-        valid_to: nil
-      })
-
-      domain = Fabricate(:domain, period: 3)
-      domain.pricelist('renew').price.amount.should == 6.1
-
-      domain = Fabricate(:domain, period: 36, period_unit: 'm')
-      domain.pricelist('renew').price.amount.should == 6.1
-
-      domain = Fabricate(:domain, period: 1095, period_unit: 'd')
-      domain.pricelist('renew').price.amount.should == 6.1
     end
 
     it 'should set pending update' do
@@ -681,7 +482,7 @@ describe Domain do
 
     invalid = [
       'a.ee', "#{'a' * 64}.ee", 'ab.eu', 'test.ab.ee', '-test.ee', '-test-.ee',
-      'test-.ee', 'te--st.ee', 'õ.pri.ee', 'test.com', 'www.ab.ee', 'test.eu', '  .ee', 'a b.ee',
+      'test-.ee', 'te--st.ee', 'õ.pri.ee', 'www.ab.ee', 'test.eu', '  .ee', 'a b.ee',
       'Ž .ee', 'test.edu.ee'
     ]
 
@@ -714,33 +515,12 @@ describe Domain do
   it 'should not create zone origin domain' do
     d = Fabricate.build(:domain, name: 'ee')
     d.save.should == false
-    d.errors.full_messages.should match_array([
-      "Data management policy violation: Domain name is blocked [name]"
-    ])
+    expect(d.errors.full_messages).to include('Data management policy violation: Domain name is blocked [name]')
 
     d = Fabricate.build(:domain, name: 'bla')
     d.save.should == false
-    d.errors.full_messages.should match_array([
-      "Domain name Domain name is invalid"
-    ])
+    expect(d.errors.full_messages).to include('Domain name Domain name is invalid')
   end
-
-  # d = Domain.new
-  # expect(d.valid?).to be false
-  # expect(d.errors.messages).to match_array({
-  # registrant: ['Registrant is missing'],
-  # admin_contacts: ['Admin contacts count must be between 1 - infinity'],
-  # nameservers: ['Nameservers count must be between 2-11'],
-  # registrar: ['Registrar is missing'],
-  # period: ['Period is not a number']
-  # })
-
-  # Setting.ns_min_count = 2
-  # Setting.ns_max_count = 7
-
-  # expect(d.valid?).to be false
-  # expect(d.errors.messages[:nameservers]).to eq(['Nameservers count must be between 2-7'])
-  # end
 
   it 'downcases domain' do
     d = Domain.new(name: 'TesT.Ee')
@@ -849,20 +629,13 @@ describe Domain do
     d.errors.full_messages.should == []
   end
 
-  it 'normalizes ns attrs' do
-    d = Fabricate(:domain)
-    d.nameservers.build(hostname: 'BLA.EXAMPLE.EE', ipv4: '   192.168.1.1', ipv6: '1080:0:0:0:8:800:200c:417a')
-    d.save
-
-    ns = d.nameservers.last
-    expect(ns.hostname).to eq('bla.example.ee')
-    expect(ns.ipv4).to eq('192.168.1.1')
-    expect(ns.ipv6).to eq('1080:0:0:0:8:800:200C:417A')
-  end
-
   it 'does not create a reserved domain' do
-    Fabricate(:reserved_domain)
-    expect(Fabricate.build(:domain, name: '1162.ee').valid?).to be false
+    Fabricate.create(:reserved_domain, name: 'test.ee')
+
+    domain = Fabricate.build(:domain, name: 'test.ee')
+    domain.validate
+
+    expect(domain.errors[:base]).to include('Required parameter missing; reserved>pw element required for reserved domains')
   end
 
   it 'validates period' do
@@ -913,8 +686,6 @@ describe Domain do
 
     context 'when saved' do
       before(:each) do
-        # Fabricate(:domain_validation_setting_group)
-        # Fabricate(:dnskeys_setting_group)
         Fabricate(:domain)
       end
 
@@ -923,6 +694,182 @@ describe Domain do
         expect(ContactVersion.count).to eq(3)
         expect(NameserverVersion.count).to eq(3)
       end
+    end
+  end
+end
+
+RSpec.describe Domain, db: false do
+  it { is_expected.to alias_attribute(:on_hold_time, :outzone_at) }
+  it { is_expected.to alias_attribute(:delete_time, :delete_at) }
+  it { is_expected.to alias_attribute(:force_delete_time, :force_delete_at) }
+  it { is_expected.to alias_attribute(:outzone_time, :outzone_at) }
+
+  describe '::expire_warning_period', db: true do
+    before :example do
+      Setting.expire_warning_period = 1
+    end
+
+    it 'returns expire warning period' do
+      expect(described_class.expire_warning_period).to eq(1.day)
+    end
+  end
+
+  describe '::redemption_grace_period', db: true do
+    before :example do
+      Setting.redemption_grace_period = 1
+    end
+
+    it 'returns redemption grace period' do
+      expect(described_class.redemption_grace_period).to eq(1.day)
+    end
+  end
+
+  describe '#set_server_hold' do
+    let(:domain) { described_class.new }
+
+    before :example do
+      travel_to Time.zone.parse('05.07.2010')
+      domain.set_server_hold
+    end
+
+    it 'sets corresponding status' do
+      expect(domain.statuses).to include(DomainStatus::SERVER_HOLD)
+    end
+
+    it 'sets :outzone_at to now' do
+      expect(domain.outzone_at).to eq(Time.zone.parse('05.07.2010'))
+    end
+  end
+
+  describe '#admin_contact_names' do
+    let(:domain) { described_class.new }
+
+    before :example do
+      expect(Contact).to receive(:names).and_return('names')
+    end
+
+    it 'returns admin contact names' do
+      expect(domain.admin_contact_names).to eq('names')
+    end
+  end
+
+  describe '#admin_contact_emails' do
+    let(:domain) { described_class.new }
+
+    before :example do
+      expect(Contact).to receive(:emails).and_return('emails')
+    end
+
+    it 'returns admin contact emails' do
+      expect(domain.admin_contact_emails).to eq('emails')
+    end
+  end
+
+  describe '#tech_contact_names' do
+    let(:domain) { described_class.new }
+
+    before :example do
+      expect(Contact).to receive(:names).and_return('names')
+    end
+
+    it 'returns technical contact names' do
+      expect(domain.tech_contact_names).to eq('names')
+    end
+  end
+
+  describe '#nameserver_hostnames' do
+    let(:domain) { described_class.new }
+
+    before :example do
+      expect(Nameserver).to receive(:hostnames).and_return('hostnames')
+    end
+
+    it 'returns name server hostnames' do
+      expect(domain.nameserver_hostnames).to eq('hostnames')
+    end
+  end
+
+  describe '#primary_contact_emails' do
+    let(:domain) { described_class.new }
+
+    before :example do
+      expect(domain).to receive(:registrant_email).and_return('registrant@test.com')
+      expect(domain).to receive(:admin_contact_emails).and_return(%w(admin.contact@test.com admin.contact@test.com))
+    end
+
+    it 'returns unique list of registrant and administrative contact emails' do
+      expect(domain.primary_contact_emails).to match_array(%w(
+                                                   registrant@test.com
+                                                   admin.contact@test.com
+                                                 ))
+    end
+  end
+
+  describe '#set_graceful_expired' do
+    let(:domain) { described_class.new }
+
+    before :example do
+      expect(described_class).to receive(:expire_warning_period).and_return(1.day)
+      expect(described_class).to receive(:redemption_grace_period).and_return(2.days)
+      expect(domain).to receive(:valid_to).and_return(Time.zone.parse('05.07.2010 10:30'))
+
+      domain.set_graceful_expired
+    end
+
+    it 'sets :outzone_at to :valid_to + expire warning period' do
+      expect(domain.outzone_at).to eq(Time.zone.parse('06.07.2010 10:30'))
+    end
+
+    it 'sets :delete_at to :outzone_at + redemption grace period' do
+      expect(domain.delete_at).to eq(Time.zone.parse('08.07.2010 10:30'))
+    end
+  end
+
+  describe '::outzone_candidates', db: true do
+    before :example do
+      travel_to Time.zone.parse('05.07.2010 00:00')
+
+      Fabricate(:zonefile_setting, origin: 'ee')
+
+      Fabricate.create(:domain, id: 1, outzone_time: Time.zone.parse('04.07.2010 23:59'))
+      Fabricate.create(:domain, id: 2, outzone_time: Time.zone.parse('05.07.2010 00:00'))
+      Fabricate.create(:domain, id: 3, outzone_time: Time.zone.parse('05.07.2010 00:01'))
+    end
+
+    it 'returns domains with outzone time in the past' do
+      expect(described_class.outzone_candidates.ids).to eq([1])
+    end
+  end
+
+  describe '::delete_candidates', db: true do
+    before :example do
+      travel_to Time.zone.parse('05.07.2010 00:00')
+
+      Fabricate(:zonefile_setting, origin: 'ee')
+
+      Fabricate.create(:domain, id: 1, delete_time: Time.zone.parse('04.07.2010 23:59'))
+      Fabricate.create(:domain, id: 2, delete_time: Time.zone.parse('05.07.2010 00:00'))
+      Fabricate.create(:domain, id: 3, delete_time: Time.zone.parse('05.07.2010 00:01'))
+    end
+
+    it 'returns domains with delete time in the past' do
+      expect(described_class.delete_candidates.ids).to eq([1])
+    end
+  end
+
+  describe '#new_registrant_email' do
+    let(:domain) { described_class.new(pending_json: { new_registrant_email: 'test@test.com' }) }
+
+    it 'returns new registrant\'s email' do
+      expect(domain.new_registrant_email).to eq('test@test.com')
+    end
+  end
+
+  describe '#new_registrant_id' do
+    let(:domain) { described_class.new(pending_json: { new_registrant_id: 1 }) }
+
+    it 'returns new registrant\'s id' do
+      expect(domain.new_registrant_id).to eq(1)
     end
   end
 end